build: Yet more simplifying of os/linux/config.mk
To further improve colocation of chipset information, definitions
related to both the chipset and the RT28xx_MODE are now stored in
the appropriate chipset definition file. Not only does this remove
information from os/linux/config.mk where it does not belong, but
by putting it in the chipset definition file, it makes it more clear
what is defined and what is supposed to be defined.

Signed-off-by: Peter Mitsis <peter.mitsis@gmail.com>